Add files via upload
I changed the model file and created a new trainer function that can handle batches manually. I also updated the feasibility checker in the utils. Unfortunately, I had to make most things dense in the playground, which is why I created the new version. I observed no training progress in the training plot, so it probably needs a double-check.